+++ resources/views/layouts/partial/footer.blade.php
@@ -7,6 +7,9 @@
           <a href="{{route('site.terms')}}" class="text-dark p-2">{{__('site.terms')}}</a>
           <a href="{{route('site.privacy')}}" class="text-dark p-2">{{__('site.privacy')}}</a>
           <a href="{{route('site.language')}}" class="text-dark p-2">{{__('site.language')}}</a>
+          @if(config_cache('instance.has_legal_notice'))
+            <a href="/site/legal-notice" class="text-dark p-2">Legal Notice</a>
+          @endif
         </p>
         <p class="text-center text-muted small mb-0">
           <span class="text-muted">© {{date('Y')}} {{config('pixelfed.domain.app')}}</span>